pyproject.toml
[tool.poetry]
name = "gerapy"
version = "0.9.13"
description = "Distributed Crawler Management Framework Based on Scrapy, Scrapyd, Scrapyd-Client, Scrapyd-API, Django and Vue.js."
authors = ["Germey <cqc@cuiqingcai.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Gerapy/Gerapy"
include = [
    "LICENSE"
]
classifiers = [
    # Trove classifiers - https://packaging.python.org/specifications/core-metadata/#metadata-classifier
    # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: Implementation :: CPython',
    'Programming Language :: Python :: Implementation :: PyPy'
]

[tool.poetry.dependencies]
python = "^3.7"
apscheduler = ">=3.5.1,<=3.7.0"
cryptography = ">=2.8"
django = ">=2.2,<3.0"
django-cors-headers = ">=3.2.0,<=3.7.0"
django-apscheduler = ">=0.3.0,<=0.6.0"
furl = ">=2.1.0"
jinja2 = ">=2.11.3"
scrapy = ">=2.7.1"
scrapy-redis = ">=0.6.8"
scrapy-splash = ">=0.7.2"
python-scrapyd-api = ">=2.1.2"
redis = ">=2.10.5"
requests = ">=2.20.0"
pymongo = ">=3.9.0"
pymysql = ">=0.7.10"
pyquery = ">=1.2.17"
beautifulsoup4 = ">=4.7.0"
djangorestframework = ">=3.11.2"
websocket = ">=0.2.1"
pyppeteer = ">=0.0.25"
uberegg = "^0.1.1"
azure-identity = "^1.13.0"
azure-keyvault-secrets = "^4.7.0"
psycopg2-binary = "^2.9.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gerapy = "gerapy.cmd:cmd"
